Fix bug in SugarCodePrinter.
- SugarCodePrinter falsely set out.length when recursing on statements.
  Now we ignore statements that are recursively visited, we may want to
  reconsider this later in case this turns out to miss a lot of names.
- SugarSuite relied on 2.11 specific implicit conversions,
  Predef.augmentString is the same for 2.11/2.12.
