+++ scalahost/integration/src/main/scala/example/Sugar.scala
@@ -3,4 +3,5 @@ package example
 class Sugar {
   List(1).map(_ + 2)
   Array.empty[Int].headOption
+  "fooo".stripPrefix("o")
 }
