+++ scalameta/tests/jvm/src/test/resources/semanticdb.expect
@@ -42,6 +42,7 @@ Names:
 [60..65): empty => _root_.scala.Array.empty(Lscala/reflect/ClassTag;)Ljava/lang/Object;.
 [66..69): Int => _root_.scala.Int#
 [71..81): headOption => _root_.scala.collection.TraversableLike#headOption()Lscala/Option;.
+[91..102): stripPrefix => _root_.scala.collection.immutable.StringLike#stripPrefix(Ljava/lang/String;)Ljava/lang/String;.
 
 Denotations:
 _root_.example. => package example
@@ -55,6 +56,7 @@ _root_.scala.Int#`<init>`()V. => primaryctor <init>: ()Int
 _root_.scala.collection.TraversableLike#headOption()Lscala/Option;. => def headOption: Option[A]
 _root_.scala.collection.immutable.List#map(Lscala/Function1;Lscala/collection/generic/CanBuildFrom;)Ljava/lang/Object;. => final def map: [B, That](f: A => B)(implicit bf: scala.collection.generic.CanBuildFrom[List[A],B,That])That
 _root_.scala.collection.immutable.List. => final object List
+_root_.scala.collection.immutable.StringLike#stripPrefix(Ljava/lang/String;)Ljava/lang/String;. => def stripPrefix: (prefix: String)String
 
 Sugars:
 [37..37): *.apply[Int]
@@ -71,3 +73,6 @@ Sugars:
 [70..70): *(((ClassTag.Int): scala.reflect.ClassTag[Int]))
   [0..1): * => _star_.
   [13..16): Int => _root_.scala.reflect.ClassTag.Int.
+[84..90): scala.Predef.augmentString(*)
+  [13..26): augmentString => _root_.scala.Predef.augmentString(Ljava/lang/String;)Ljava/lang/String;.
+  [27..28): * => _star_.
